# swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Bonza Bike Tours REST API",
		"description": "REST API description for the Bonza Bike Tours BMS dashboard.",
		"contact": {
			"name": "API Support",
			"email": "alexander.shumilov@toptal.com"
		},
		"version": "1.2"
	},
	"servers": [{
			"url": "bonzabiketours-api.herokuapp.com/api/v1",
			"description": "Heroku server - CI"
		},
		{
			"url": "bms.bonzabiketours.alexshumilov.ru:3001/api/v1",
			"description": "Staging server"
		},
		{
			"url": "bms.bonzabiketours.alexshumilov.ru:3002/api/v1",
			"description": "Development server"
		}
	],
	"security": [{
		"api_key": []
	}],
...
	"paths": {
...
		"/bookingPartners": {
			"get": {
				"tags": [
					"bookingPartners"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Returns booking partners according to filters",
				"parameters": [{
						"name": "bookingPartnerId",
						"description": "Booking partner identifier",
						"in": "query",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "name",
						"description": "Name filter",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "reservationConfirmEmail",
						"description": "Reservation confirm email filter",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "commissionLevel",
						"description": "Commission level",
						"in": "query",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "sortBy",
						"description": "Sorted column",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"$ref": "#/components/parameters/limit"
					},
					{
						"$ref": "#/components/parameters/page"
					}
				],
				"responses": {
					"200": {
						"description": "Booking partners search result",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/BookingPartner"
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			},
			"post": {
				"tags": [
					"bookingPartners"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Create booking partner",
				"requestBody": {
					"description": "Booking partner description",
					"content": {
						"application/json": {
							"schema": {
								"title": "",
								"properties": {
									"reservationConfirmEmail": {
										"type": "string"
									},
									"paymentDueTerm": {
										"type": "string"
									},
									"balanceDueDays": {
										"type": "string"
									},
									"name": {
										"type": "string"
									},
									"myob": {
										"type": "string"
									},
									"commissionLevel": {
										"type": "integer"
									},
									"contactFirstname": {
										"type": "string"
									},
									"contactLastname": {
										"type": "string"
									},
									"contactEmail": {
										"type": "string"
									},
									"address": {
										"type": "string"
									},
									"city": {
										"type": "string"
									},
									"state": {
										"type": "string"
									},
									"postcode": {
										"type": "string"
									},
									"country": {
										"type": "string"
									},
									"phone": {
										"type": "string"
									},
									"fax": {
										"type": "string"
									},
									"website": {
										"type": "string"
									},
									"emailConfirmation": {
										"type": "string"
									},
									"reservationPhone": {
										"type": "string"
									},
									"reservationEmail": {
										"type": "string"
									},
									"blackoutPhone": {
										"type": "string"
									},
									"blackoutEmail": {
										"type": "string"
									},
									"paymentMethod": {
										"type": "string"
									},
									"adultPrice": {
										"type": "string"
									},
									"childPrice": {
										"type": "string"
									},
									"familyPrice": {
										"type": "string"
									},
									"familyAdditionalRiderPrice": {
										"type": "string"
									},
									"layout": {
										"type": "string"
									},
									"printedForm": {
										"type": "string"
									},
									"bookingPartnerNotes": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Created booking partner identifier",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APICreatedEntityResponse"
								}
							}
						}
					}
				}
			}
		},
		"/bookingPartners/{bookingPartnerId}": {
			"get": {
				"tags": [
					"bookingPartners"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Return booking partner",
				"parameters": [{
					"name": "bookingPartnerId",
					"description": "Booking partner identifier",
					"in": "path",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"responses": {
					"200": {
						"description": "Booking partner",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/BookingPartner"
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			},
			"put": {
				"tags": [
					"bookingPartners"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Update booking partner",
				"parameters": [{
					"name": "bookingPartnerId",
					"description": "Booking partner identifier",
					"in": "path",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"requestBody": {
					"description": "Booking partner description",
					"content": {
						"application/json": {
							"schema": {
								"title": "",
								"properties": {
									"reservationConfirmEmail": {
										"type": "string"
									},
									"paymentDueTerm": {
										"type": "string"
									},
									"balanceDueDays": {
										"type": "string"
									},
									"name": {
										"type": "string"
									},
									"myob": {
										"type": "string"
									},
									"commissionLevel": {
										"type": "integer"
									},
									"contactFirstname": {
										"type": "string"
									},
									"contactLastname": {
										"type": "string"
									},
									"contactEmail": {
										"type": "string"
									},
									"address": {
										"type": "string"
									},
									"city": {
										"type": "string"
									},
									"state": {
										"type": "string"
									},
									"postcode": {
										"type": "string"
									},
									"country": {
										"type": "string"
									},
									"phone": {
										"type": "string"
									},
									"fax": {
										"type": "string"
									},
									"website": {
										"type": "string"
									},
									"emailConfirmation": {
										"type": "string"
									},
									"reservationPhone": {
										"type": "string"
									},
									"reservationEmail": {
										"type": "string"
									},
									"blackoutPhone": {
										"type": "string"
									},
									"blackoutEmail": {
										"type": "string"
									},
									"paymentMethod": {
										"type": "string"
									},
									"adultPrice": {
										"type": "string"
									},
									"childPrice": {
										"type": "string"
									},
									"familyPrice": {
										"type": "string"
									},
									"familyAdditionalRiderPrice": {
										"type": "string"
									},
									"layout": {
										"type": "string"
									},
									"printedForm": {
										"type": "string"
									},
									"bookingPartnerNotes": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Operation status",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APISuccessResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			},
			"delete": {
				"tags": [
					"bookingPartners"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Delete booking partner",
				"parameters": [{
					"name": "bookingPartnerId",
					"description": "Booking partner identifier",
					"in": "path",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"responses": {
					"200": {
						"description": "Operation status",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/responses/APISuccessResponse"
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			}
		},
		"/bookingPartners/{bookingPartnerId}/commission": {
			"get": {
				"tags": [
					"bookingPartners"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Booking partner commission",
				"parameters": [{
					"name": "bookingPartnerId",
					"description": "Booking partner identifier",
					"in": "path",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"responses": {
					"200": {
						"description": "Commission information",
						"content": {
							"application/json": {
								"schema": {
									"title": "",
									"properties": {
										"commission": {
											"type": "string"
										},
										"email": {
											"type": "string"
										},
										"bookingRefID": {
											"type": "string"
										},
										"bookingPartnerNotes": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			}
		},
		"/bookingPartners/{bookingPartnerId}/purchases": {
			"get": {
				"tags": [
					"bookingPartners"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Booking partner purchases",
				"parameters": [{
						"name": "bookingPartnerId",
						"description": "Booking partner identifier",
						"in": "path",
						"schema": {
							"type": "integer"
						},
						"required": true
					},
					{
						"$ref": "#/components/parameters/limit"
					},
					{
						"$ref": "#/components/parameters/page"
					}
				],
				"responses": {
					"200": {
						"description": "Booking partner purchases",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/BookingPartnerPurchase"
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			}
		},
...
		"/broadcasts": {
			"get": {
				"tags": [
					"broadcasts"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Broadcast emails",
				"parameters": [{
						"name": "from",
						"description": "Sent date from",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"required": true
					},
					{
						"name": "to",
						"description": "Sent date to",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"required": true
					},
					{
						"name": "emailTo",
						"description": "Recipient",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "subject",
						"description": "Subject",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "userId",
						"description": "User identifier",
						"in": "query",
						"schema": {
							"type": "integer"
						}
					},
					{
						"name": "travelagency",
						"description": "Travel agency",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"$ref": "#/components/parameters/limit"
					},
					{
						"$ref": "#/components/parameters/page"
					}
				],
				"responses": {
					"200": {
						"description": "List of emails",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/BroadcastEmail"
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": [
					"broadcasts"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Create broadcast email",
				"requestBody": {
					"description": "Broadcast email description",
					"content": {
						"application/json": {
							"schema": {
								"title": "",
								"properties": {
									"templateCode": {
										"type": "string"
									},
									"templateCode2": {
										"type": "string"
									},
									"subject": {
										"type": "string"
									},
									"body": {
										"type": "string"
									},
									"internalNotes": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Operation status",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APICreatedEntityResponse"
								}
							}
						}
					}
				}
			}
		},
		"/broadcasts/{emailId}": {
			"put": {
				"tags": [
					"broadcasts"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Update broadcast email",
				"parameters": [{
					"name": "emailId",
					"description": "Broadcast email identifier",
					"in": "path",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"requestBody": {
					"description": "Broadcast description",
					"content": {
						"application/json": {
							"schema": {
								"title": "",
								"properties": {
									"emailID": {
										"type": "string"
									},
									"sendFrom": {
										"type": "string"
									},
									"subject": {
										"type": "string"
									},
									"message": {
										"type": "string"
									},
									"internalNotes": {
										"type": "string"
									},
									"setupDate": {
										"type": "string"
									},
									"totalSent": {
										"type": "integer"
									},
									"setupUserID": {
										"type": "integer"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Operation status",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APISuccessResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			},
			"delete": {
				"tags": [
					"broadcasts"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Delete broadcast email",
				"parameters": [{
					"name": "emailId",
					"description": "Broadcast email identifier",
					"in": "path",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"responses": {
					"200": {
						"description": "Operation status",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APISuccessResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			}
		},
		"/broadcasts/{emailId}/send": {
			"post": {
				"tags": [
					"broadcasts"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Send broadcast email",
				"parameters": [{
					"name": "emailId",
					"description": "Broadcast email identifier",
					"in": "path",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"responses": {
					"200": {
						"description": "Operation status",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APISuccessResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			}
		},
...
		"/broadcasts/attachments/{attachmentId}": {
			"delete": {
				"tags": [
					"broadcasts"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Delete attachment from the broadcast email",
				"parameters": [{
					"name": "attachmentId",
					"description": "Attachment identifier",
					"in": "path",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"responses": {
					"200": {
						"description": "Operation status",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APICreatedEntityResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			}
		},
...
	},
	"components": {
		"responses": {
...
			"APISuccessResponse": {
				"description": "Operation was completed successfully",
				"content": {
					"application/json": {
						"schema": {
							"properties": {
								"status": {
									"type": "string"
								},
								"message": {
									"type": "string"
								}
							}
						}
					}
				}
			},
...
			"APIUnauthorizedResponse": {
				"description": "Authorize in order to access the resource",
				"content": {
					"application/json": {
						"schema": {
							"properties": {
								"status": {
									"type": "string"
								},
								"errorType": {
									"type": "string"
								},
								"errorMessage": {
									"type": "string"
								}
							}
						}
					}
				}
			},
			"APINotFoundResponse": {
				"description": "Requested resource was not found",
				"content": {
					"application/json": {
						"schema": {
							"properties": {
								"status": {
									"type": "string"
								},
								"errorType": {
									"type": "string"
								},
								"errorMessage": {
									"type": "string"
								}
							}
						}
					}
				}
			},
			"APICreatedEntityResponse": {
				"description": "Returning identifier of created entity",
				"content": {
					"application/json": {
						"schema": {
							"properties": {
								"id": {
									"type": "integer"
								}
							}
						}
					}
				}
			}
		},
		"parameters": {
...
			"limit": {
				"name": "limit",
				"in": "query",
				"description": "Specifies maximum number of fetched items",
				"schema": {
					"type": "integer"
				}
			},
			"page": {
				"name": "page",
				"in": "query",
				"description": "Specifies page of fetched items (offsetting)",
				"schema": {
					"type": "integer"
				}
			}
		},
		"schemas": {
...
			"BroadcastEmail": {
				"properties": {
					"id": {
						"type": "integer"
					}
				}
			},
...
			"BookingPartner": {
				"properties": {
					"bookingPartnerId": {
						"type": "integer"
					},
					"name": {
						"type": "string"
					},
					"myob": {
						"type": "integer"
					},
					"reservationConfirmEmail": {
						"type": "string"
					},
					"commissionLevel": {
						"type": "integer"
					},
					"paymentMethod": {
						"type": "string"
					},
					"paymentDueTerm": {
						"type": "integer"
					}
				}
			},
...
			"BookingPartnerPurchase": {
				"properties": {
					"status": {
						"type": "string"
					},
					"tourDate": {
						"type": "string"
					},
					"travelerLastname": {
						"type": "string"
					},
					"productName": {
						"type": "string"
					},
					"totalGuest": {
						"type": "integer"
					},
					"outstanding": {
						"type": "string"
					},
					"bookingPartnerName": {
						"type": "string"
					},
					"firstname": {
						"type": "string"
					},
					"lastname": {
						"type": "string"
					},
					"noOfAdult": {
						"type": "integer"
					},
					"noOfChildren": {
						"type": "integer"
					},
					"noOfFamilyGroups": {
						"type": "integer"
					},
					"noOfAdditionals": {
						"type": "integer"
					},
					"purchaseDate": {
						"type": "string"
					}
				}
			},
...
		}
	}
}